💄 Nuevo diseño de historias
